feat: Migrate to net10.0. working on improving the code, async implementation, add services.

+++ MyULibraryBackend/MyULibraryBackend/Dtos/RoleDto.cs
@@ -7,7 +7,7 @@ namespace MyULibraryBackend.Dtos
 {
     public class RoleDto
     {
-        public int IdRole { get; set; }
+        public long IdRole { get; set; }
         public string RoleName { get; set; }
     }
 }
